Time: 52 ms (86.05%) | Memory: 68.6 MB (38.37%) - LeetSync

+++ 3299-find-the-maximum-number-of-elements-in-subset/find-the-maximum-number-of-elements-in-subset.java
@@ -23,14 +23,14 @@ class Solution {
                 continue;
 
             int currentLength = 0;
-            long current = num;
+            int current = num;
 
-            while (map.getOrDefault((int) current, 0) >= 2) {
+            while (map.getOrDefault( current, 0) >= 2) {
                 currentLength += 2;
                 current *= current;
             }
 
-            if (map.getOrDefault((int) current, 0) > 0) {
+            if (map.getOrDefault( current, 0) > 0) {
                 currentLength += 1;
             } else {
                 currentLength -= 1;
